Move metadata and constants into /src/shared
