+++ models/ambulance.js
@@ -1,6 +1,6 @@
 const { Model } = require('sequelize');
-const metadata = require('../src/metadata/ambulance');
-const convertToSequelizeField = require('../src/metadata/convertToSequelizeField');
+const metadata = require('../src/shared/metadata/ambulance');
+const convertToSequelizeField = require('../src/shared/metadata/convertToSequelizeField');
 
 module.exports = (sequelize) => {
   class Ambulance extends Model {
