+++ models/hospital.js
@@ -1,8 +1,8 @@
 const _ = require('lodash');
 const { Model } = require('sequelize');
-const { DeliveryStatus } = require('../src/constants');
-const metadata = require('../src/metadata/hospital');
-const convertToSequelizeField = require('../src/metadata/convertToSequelizeField');
+const { DeliveryStatus } = require('../src/shared/constants');
+const metadata = require('../src/shared/metadata/hospital');
+const convertToSequelizeField = require('../src/shared/metadata/convertToSequelizeField');
 
 module.exports = (sequelize) => {
   class Hospital extends Model {
