+++ models/hospitalStatusUpdate.js
@@ -1,8 +1,8 @@
 const _ = require('lodash');
 const { Model, Op } = require('sequelize');
-const { DeliveryStatus } = require('../src/constants');
-const metadata = require('../src/metadata/hospitalStatusUpdate');
-const convertToSequelizeField = require('../src/metadata/convertToSequelizeField');
+const { DeliveryStatus } = require('../src/shared/constants');
+const metadata = require('../src/shared/metadata/hospitalStatusUpdate');
+const convertToSequelizeField = require('../src/shared/metadata/convertToSequelizeField');
 
 module.exports = (sequelize) => {
   class HospitalStatusUpdate extends Model {
