+++ models/user.js
@@ -1,8 +1,8 @@
 const bcrypt = require('bcrypt');
 const _ = require('lodash');
 const { Model } = require('sequelize');
-const metadata = require('../src/metadata/user');
-const convertToSequelizeField = require('../src/metadata/convertToSequelizeField');
+const metadata = require('../src/shared/metadata/user');
+const convertToSequelizeField = require('../src/shared/metadata/convertToSequelizeField');
 
 const SALT_ROUNDS = 10;
