+++ routes/api/ringdowns.js
@@ -5,7 +5,7 @@ const _ = require('lodash');
 const middleware = require('../../auth/middleware');
 const models = require('../../models');
 const { dispatchRingdownUpdate } = require('../../wss');
-const { DeliveryStatus } = require('../../src/constants');
+const { DeliveryStatus } = require('../../src/shared/constants');
 
 const { setPaginationHeaders } = require('../helpers');
