+++ src/shared/metadata/hospital.js
